+++ java/coverage-tests/src/test/java/coverage/drivers/browsers/ChromeEmulatorBuilder.java
@@ -11,14 +11,20 @@ import org.openqa.selenium.remote.RemoteWebDriver;
 
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.List;
 
 public class ChromeEmulatorBuilder implements DeviceBuilder {
 
     public void browser(String browser) {}
 
-    public WebDriver build(boolean headless, boolean legacy, boolean executionGrid) throws MalformedURLException {
+    public WebDriver build(boolean headless, boolean legacy, boolean executionGrid, List<String> args) throws MalformedURLException {
         ChromeOptions options = CapabilitiesHelper.getAndroid8ChromeEmulator(headless);
         options.addArguments("--remote-allow-origins=*");
+        if (args != null) {
+            for (String arg : args) {
+                options.addArguments(arg);
+            }
+        }
         if (GlobalSetup.CI) {
             options.addArguments("--no-sandbox", "--disable-gpu");
         }