+++ java/eyes.sdk.core/src/main/java/com/applitools/eyes/settings/GetResultsSettings.java
@@ -6,10 +6,17 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 public class GetResultsSettings {
 
     private Boolean throwErr;
+    private Boolean removeDuplicateTests;
 
-    public GetResultsSettings(Boolean throwErr) {
-        this.throwErr = throwErr;
+    public GetResultsSettings(boolean shouldThrowException) {
+        this.throwErr = shouldThrowException;
     }
+
+    public GetResultsSettings(Boolean throwErr, Boolean removeDuplicateTests) {
+        this.throwErr = throwErr;
+        this.removeDuplicateTests = removeDuplicateTests;
+    }
+
 
     public Boolean getThrowErr() {
         return throwErr;
@@ -17,5 +24,13 @@ public class GetResultsSettings {
 
     public void setThrowErr(Boolean throwErr) {
         this.throwErr = throwErr;
+    }
+
+    public Boolean getRemoveDuplicateTests() {
+        return removeDuplicateTests;
+    }
+
+    public void setRemoveDuplicateTests(Boolean removeDuplicateTests) {
+        this.removeDuplicateTests = removeDuplicateTests;
     }
 }