+++ java/eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/Eyes.java
@@ -1930,7 +1930,8 @@ public class Eyes implements IEyesBase {
     }
 
     private Map<String, List<Region>> locateDto(ITargetDto target, VisualLocatorSettingsDto settings, ConfigurationDto config) {
-        return commandExecutor.locate(target, settings, config);
+        Map<String, List<RegionDto>> response = commandExecutor.locate(target, settings, config);
+        return LocateMapper.toLocateFromDto(response);
     }
 
     // in v3 this was changed to 'locateText'.